Rename countries/index.js to countries/utils.js
