+++ src/main/tray/menu-builder.js
@@ -17,8 +17,8 @@
 
 // @flow
 
-import type { Country } from '../../app/countries/index'
-import { getCountryLabel } from '../../app/countries/index'
+import type { Country } from '../../app/countries/utils'
+import { getCountryLabel } from '../../app/countries/utils'
 import ConnectionStatusEnum from 'mysterium-tequilapi/lib/dto/connection-status-enum'
 import type { ConnectionStatus } from 'mysterium-tequilapi/lib/dto/connection-status-enum'
 import TrayMenu from './menu'
